requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2
python-decouple==3.8
Pillow==10.1.0